clean フォルダ構成を整理しました.
