+++ Sources/GameWidget/Commons/StructuringTools/NodeWidget/NodeContext.swift
