+++ Sources/GameWidget/Commons/StructuringTools/NodeWidget/NodeWidget.swift
