+++ Sources/GameWidget/Modifiers/ForAllNode.swift
